move TokenBlaster to its own file
